Добавление компонента для автодополнения полей.
